requirements.txt
asgiref==3.2.7
certifi==2019.3.9
chardet==3.0.4
cleanco==1.3
Django==2.1.5
django-dynamic-fixture==2.0.0
django-extensions==2.1.6
django-redis==4.10.0
djangodeletes==0.43
djangorestframework==3.9.2
fuzzywuzzy==0.17.0
gprof2dot==2017.9.19
idna==2.8
numpy==1.18.3
pandas==0.24.2
psycopg2-binary==2.8.3
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.1
urllib3==1.24.1
uWSGI==2.0.18
